feat(combat): add damage type/armor type modifier matrix (#62)
Add a complete combat damage modifier system with:
- DamageType enum (Melee, Ranged, Siege, Magic, True) and ArmorType enum
  (InfantryLight, InfantryHeavy, Ranged, Cavalry, Structure, StructureHuman)
- CombatModifiers static class with 5x6 modifier matrix and
  CalculateFinalDamage pipeline (type modifier * height * crystal * defense)
- Defense component changed to int fields (Melee, Ranged, Siege, Magic)
  with diminishing-returns formula: 1 - def/(def+100)
- DmgType field added to Projectile struct for damage type propagation
- MeleeCombatSystem and ProjectileSystem integrated with full modifier pipeline
- All projectile-spawning systems (Ranged, Building, Veilstinger, Godsplinter)
  propagate DamageType through DmgType field
- All 10 units tagged with DamageTypeData, ArmorTypeData, and Defense values
- All 13 buildings tagged with ArmorTypeData (and DamageTypeData where they
  have ranged attacks: Hall, FiendstoneKeep, CrystalTurretNode)

+++ Assets/Scripts/Core/Components/BuildingComponents.cs
@@ -170,21 +170,23 @@ public struct RepairOrder : IComponentData
 /// </summary>
 public struct DeferredDefense : IComponentData
 {
-    public float Melee;
-    public float Ranged;
-    public float Siege;
-    public float Magic;
+    public int Melee;
+    public int Ranged;
+    public int Siege;
+    public int Magic;
 }
 
 /// <summary>
-/// Defensive stats for completed buildings.
+/// Defensive stats for buildings and units.
+/// Each field reduces incoming damage of that type via diminishing-returns formula:
+///   reduction = defense / (defense + 100)
 /// </summary>
 public struct Defense : IComponentData
 {
-    public float Melee;
-    public float Ranged;
-    public float Siege;
-    public float Magic;
+    public int Melee;
+    public int Ranged;
+    public int Siege;
+    public int Magic;
 }
 
 // ==================== Training System ====================
